+++ src/shapes/Object/FabricObject.spec.ts
@@ -1,6 +1,6 @@
 import { FabricObject } from './FabricObject';
 
-import { describe, expect, it, vi } from 'vitest';
+import { describe, expect, it } from 'vitest';
 
 describe('FabricObject', () => {
   it('setCoords should calculate control coords only if canvas ref is set', () => {
@@ -10,7 +10,8 @@ describe('FabricObject', () => {
     object.setCoords();
     expect(object.aCoords).toBeDefined();
     expect(object.oCoords).toBeUndefined();
-    object.canvas = vi.fn();
+    // @ts-expect-error -- mock canvas
+    object.canvas = { getZoom: () => 1 };
     object.setCoords();
     expect(object.aCoords).toBeDefined();
     expect(object.oCoords).toBeDefined();