+++ src/util/misc/matrix.ts
@@ -109,6 +109,22 @@ export const calcPlaneRotation = ([a, b]: TMat2D) =>
   Math.atan2(b, a) as TRadian;
 
 /**
+ * Returns the uniform scale (zoom) magnitude of a 2D affine matrix,
+ * computed as the length of the image of the unit X basis vector.
+ */
+export const calcPlaneZoom = ([a, b]: TMat2D) => Math.sqrt(a * a + b * b);
+
+/**
+ * Returns the Y-axis scale magnitude of a 2D affine matrix,
+ * computed as the length of the image of the unit Y basis vector.
+ * We do not support non uniform zooming on the viewport but to make things work
+ * we need to have a function that can return the value on Y axis.
+ * Bug reports and features will be planned with zoom being just a number
+ * that is the same for both axis
+ */
+export const calcPlaneScaleY = ([, , c, d]: TMat2D) => Math.sqrt(c * c + d * d);
+
+/**
  * Decomposes standard 2x3 matrix into transform components
  * @param  {TMat2D} a transformMatrix
  * @return {Object} Components of transform